backend/requirements.txt.txt
fastapi
uvicorn
paho-mqtt
influxdb-client
